+++ apps/web/src/app/[orgShortcode]/settings/org/mail/addresses/page.tsx
@@ -1,11 +1,13 @@
 'use client';
 
 import { useGlobalStore } from '@/src/providers/global-store-provider';
+import ExternalEmailModal from './_components/external-email-modal';
 import { AddEmailModal } from './_components/add-address-modal';
 import { DataTable } from '@/src/components/shared/table';
 import { SpinnerGap } from '@phosphor-icons/react';
 import { columns } from './_components/columns';
 import { platform } from '@/src/lib/trpc';
+import { useState } from 'react';
 
 export default function Page() {
   const orgShortcode = useGlobalStore((state) => state.currentOrg.shortcode);
@@ -14,6 +16,9 @@ export default function Page() {
       orgShortcode
     });
 
+  const [addExternalModalOpen, setAddExternalModalOpen] =
+    useState<boolean>(false);
+
   return (
     <div className="flex w-full flex-col gap-2 p-4">
       <div className="flex w-full justify-between">
@@ -21,7 +26,11 @@ export default function Page() {
           <h1 className="font-display text-3xl leading-5">Addresses</h1>
           <div>Manage Your Org Email Addresses</div>
         </div>
-        <AddEmailModal />
+        <AddEmailModal addExternalModalOpen={setAddExternalModalOpen} />
+        <ExternalEmailModal
+          open={addExternalModalOpen}
+          addExternalModalOpen={setAddExternalModalOpen}
+        />
       </div>
       {isLoading && (
         <div className="flex w-full justify-center gap-2 text-center font-bold">
